don't promote precision (float->double)

+++ src/util.c
@@ -42,8 +42,8 @@ uint64_t hamming_distance (uint64_t hash1, uint64_t hash2) {
 }
 
 int compare_floats (const void* a, const void* b) {
-  double arg1 = *(const float*)a;
-  double arg2 = *(const float*)b;
+  float arg1 = *(const float*)a;
+  float arg2 = *(const float*)b;
 
   if (arg1 < arg2) {
     return -1;
